requirements.txt
cycler==0.10.0
keras==2.6.0
kiwisolver==1.3.1
matplotlib==3.4.3
numpy==1.21.1
opencv-contrib-python==4.5.3.56
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
